Add a file for an automatic model running
This jupyter script creates folders with model runs for different
dividends payout ratios experiment.

You should set a path where script will save data and a path to a
file with an initial state of the model. Also you need to set a
decired number of years of a simulation. Also you can change
a frequency of saves of output data files. By default this value
is 60 days.

Signed-off-by: Nikita Sobolev <niksob97@gmail.com>